Remove use of context-wrapping object to expressing everything in terms of certs.
Prior to this change, a per-connection context was used either with system default authorities or ones given by the requester. This was both wasteful, and problematic as there is no consistent way to get authorities across all systems. Now the context is abtracted away and presented as a set of authorities, and intent on how a connection should be verified (or not) is explicit.

#rb per.larsson
#rnx

[CL 33652943 by martin ridgers in ue5-main branch]