Added the ability for Mass processors/observers to provide a description.
The Mass processor is updated with a virtual function that allows a description to be provided as key/value string pairs. When available the Mass Debugger will show these, otherwise the Mass Debugger will look the same as before. This change was added because TEDS has additional information that it appended to the name of the processor. This is increasingly difficult to read in the Mass Debugger and leads to other problems when needing to reference the name of a processor.

#rb Mieszko.Zielinski, mikko.mononen
[RN] If the debug output description provided by a Mass processor now has more than the processor name it it, the description will be shown in the Mass Debugger.

[CL 33670251 by ronald koppers in ue5-main branch]