OpenImageDenoise: Update to OIDN 2.2.2
This new version is slightly faster on the default settings. Note the binary was built without GPU support for simplicity for now. I also patched the library to not include the weights that our plugins never make use of, saving around 10Mb from the main dll. A follow up CL will investigate the new features of this release such as alpha denoising and mixed format denoising (half vs. float).

Update the TPS and license files to match the updated version.

#rb aleksander.netzel

[CL 33672744 by chris kulla in ue5-main branch]