Displaying categories for custom notify class properties in the details panel. Moving category text code from DetailCategoryBuilderImpl to ObjectEditorUtils.
All properties on the notify class will now show in their appropriate categories in the detail panel when selecting an animation notify in the animation editor. The categories will be placed as subgroups of the "Anim Notify" group. Any properties without a category will be placed directly under the "Anim Notify" group.

Properties that explicitly specify a category of "AnimNotify" will be added to the "Anim Notify" category directly as if they did not have the "AnimNotify" prefix to prevent showing an extra nested category.

[REVIEW]
#tests viewing and editing anim notifies with different categories and advanced properties in the animation editor
#rb dan.oconnor, logan.buchy

[CL 33675642 by spencer melnick in ue5-main branch]