Enable Nanite fallback rasterization for displaced clusters in shadow permutations.
"Displacement Fade" was previously added to materials to allow the system to subtly fade out and then disable tessellation of whole Nanite static meshes (and individual clusters on Landscape). However, this ability was disallowed in shadow, as self-shadowing artifacts could occur in areas where non-displaced shadow geometry shadows displaced geometry in the main view.

This change adds the ability for Nanite rasterization in shadow to uniformly apply minimum displacement to fallback clusters in non-tessellated vertex processing permutations. This allows for faster, non-tessellated fallback rasterization of displaced materials while continuing to prevent the self-shadowing artifacts.

There are a couple known issues with the approach taken to achieve this. In order to not incur any additional cost to normal fixed function shadow rasterization, an additional permutation has been added to the vertex and compute rasterization shaders for the default material only. No additional permutation is necessary for displacement fallback raster bins that are still programmable, and the additional logic is merely added to the rasterization shaders without the NANITE_TESSELLATION permutation. Displacement fallback bins must now also maintain duplicate displacement scaling in the raster bin meta data (as opposed to leaving it to default), which can potentially further worsen the number of additional raster bins rendered when using many unique overrides to these values among child material instances.

#rb Brian.Karis
[FYI] rune.stubbe, graham.wihlidal, ola.olsson, andrew.lauritzen

[CL 33691589 by jamie hayes in ue5-main branch]