In SDockingTabStack::OpenPersistentTab consider a persistent opened tab as a closed tab if it is absent from the live tabs.
This fixes an issue where persistent opened tabs could accumulate over time.  This situation can occur when the editor process is terminated or if the layout is saved during the shutdown sequence before an editor mode is deactivated.

In FTabManager::RestoreArea_Helper allow the call to SpawnTab to return nullptr instead of creating an empty "unrecognized tab".

This case can happen after one of the shutdown sequences described above when we attempt to restore a persistent opened tab but there is no currently registered tab spawner for it.  Now, we no longer present the unhelpful "unrecognized tab" but we allow the persistent opened tab to be treated as a closed tab and thus opened successfully at a later point in time.

#jira UE-211194, UE-210631, UE-120304
#rb aditya.ravichandran, ben.zeigler

[CL 33693099 by ross smith2 in ue5-main branch]