Disabling the Windows Compiler warning C6393.
The Warning:
warning C6393: A lookup table of size 365 is not sufficient to handle leap years.

This is a warning that goes off if any static array is 365 members long.  The intent of the warning is to stop common misuse of static arrays to look up days of the year.

In practice someone hitting this could just disable the warning locally to approve the array and remove the warning.  In our case though the code generation step can easily generate arrays 365 members long and there is no (good) way to fix it.

So turning it off globally makes the most sense for us.


#rb Marc.Audy, Rob.Cannaday

[CL 33697690 by thomas mauer in ue5-main branch]