Cameras: better manage loose nodes in the camera rig graph editor
Root objects like camera rigs can now hold (for editor builds) a list of all their children objects, including the "loose" ones that would otherwise not be connected to them. A new object interface provides support for that, and the object-tree graph editor uses it to add/remove objects when corresponding nodes are added/removed.

This fixes the problem of loose camera nodes disappearing from the graph, and the problem of deleted camera nodes reappearing in it. In one case it was because these nodes were GC'ed eventually, and in the other case it was because the nodes were kept alive by undo transactions.

[CL 33726067 by ludovic chabant in ue5-main branch]