InteractiveToolsFramework: Separate the BaseBrushTool's BrushAdjusterInputBehavior click drag handling into its own behavior.
Previously BrushAdjusterInputBehavior was used to set a toggle that would be read during the On*Drag callbacks to differentiate between a "brush stroke" and a "brush adjustment". This was error prone for both new and old tools that derived off of BaseBrushTool as it broke or was easy to break assumptions made in the tool's drag callbacks. With this separation, OnBeginDrag/OnUpdateDrag/OnEndDrag will only be called for brush strokes.

Additional minor fixes:
- Disabled the BrushAdjusterBehavior for SeamSculpt.
- Fixed initial BrushAdjuster HUD display prior to first drag.

#rb Jimmy.Andrews
#jira UE-213223

[CL 33797602 by lonnie li in ue5-main branch]