Hook into the GC to make sure objects in TEDS are properly cleaned up.
This change hookst TEDS Compatibility into the garbage collection so objects are removed from TEDS properly. The immediate effect is that several places where objects are tracked can be removed in the future and there's now a stronger guarantee that objects are still accessible in a OnRemove observer. Three new console commands were added for testing purpose:
1. TEDS.Feature.IntegrateWithGC - Enables the integration.
2. TEDS.Feature.ActorsClearedByGC - Disables places where actors were explicitly removed and lets the GC take care of it.
3. TEDS.Feature.EntitiesClearedByGC - Same as the previous but for entities.
These options are currently all off by default. Turning the actor or entity options on will automatically turn on the integration as well.

In order to let the editor know what features TEDS support the 3 main interfaces have been extended with functions to retrieve extension names. Right now this allows turning on/off the actor/entity support, but this can in the future be used for other extensions as well, in particular extension we want to keep in-house or only supply to specific customers. The console command "TEDS.Debug.ListExtensions" was added to discover the available extension.

To address an issue with TSAN, there's now a check for any undo operation happening on the game thread.

#jira UE-215588 UE-215094
#rb brooke.hubert, logan.buchy
#rnx

[CL 33840437 by ronald koppers in ue5-main branch]