Explicitly handle out-of-memory errors from embree calls
When an embree call fails because it runs out of memory, we don't currently do much except locally return. As evidenced in UEFN cooks, this leaves the cook in an uncertain state that will fail later and isn't properly reported as a memory issue. We can instead invoke the existing codepath for reporting out-of-memory issues (which will also stop the process).

#rb Laura.Hermanns

[CL 33857109 by sebastian schoner in ue5-main branch]