Cameras: fix copy/paste
Previously copy/paste was implemented by copying the graph nodes, and post-processing them by duplicating their underlying objects and patching up their references. This proved to be more complicated than it was worth, and wasn't solid when copy/pasting between different camera rigs.

This new implementation does the opposite: it duplicates the underlying objects, and then recreates brand new nodes from the duplicates. This is much better since the whole concept of the object-tree-graph module is to use the objects as the ground truth, and the graph nodes as something that can always be rebuilt from scratch.

[CL 33916627 by ludovic chabant in ue5-main branch]