Cameras: improve how object-tree-graphs keep track of loose objects
This change refactors how the "connectable objects" APIs on IObjectTreeGraphRootObject work. This clarifies the responsibilities between classes between those that affect the underlying data objects and those that only care about the graph nodes. It also makes it possible to have loose objects on multiple graphs related to the same asset, such as the node-tree and transitions graphs for camera rigs.

[CL 33916743 by ludovic chabant in ue5-main branch]