[Motion Design] Rundown Server: Fix references held by the rundown server that prevents deleting rundown assets.
1 - Rundown Server holding reference preventing assets from being deleted

The rundown server is considered an editor and should hold a references to what is being edited as a principle (similar to FAssetEditorToolkit::FGCEditingObjects).

The rundown server maintains contextual caches for each of the commands. Contexts are split in editing and playback categories, each requiring different contextual resources (i.e. either managed instances for page editing through RC or playback instances for actual playback). The rundown server will try to hold on to those contexts in order to save time on successive requests, but it shouldn't prevent assets from being deleted if it makes sense to do so. Fortunately, FEditorDelegates::OnAssetsPreDelete can be used to have the rundown server flush it's command contexts when an asset is about to be deleted. This works both from the content browser and other commands deleting assets through ObjectTools asset and object deleting functions.

In order to make the editing and playback contexts easier to "flush", some of the code around what was FRundownEditCommandData and FRundownPlaybackCommandData has been reworked. I found it more appropriate to call those "contexts" and move all the common "rundown server to rundown" delegates binding to a separate object: FRundownEntry, that has it's own separate cache and can be shared amongts the command contexts. Given that we have independent command contexts and, in the future, it is to be expected that there will be more (i.e. one editing context per client for instance), we really want to decouple the per-rundown objects from the per-command context ones. The FRundowEntry manages the rundown server's delegates in the rundown, while the command contexts manage the associated resources. I think it is clearer that way.

1.1 - Playback Manager Delegates

Moved the Rundown Server to Playback Manager's delegate binding to one place at init of the rundown server. It is less convoluted (imho) and there will potentially be more delegates to listen to (i.e. OnPlaybackInstanceInvalidated, OnLocalPlaybackAssetRemoved) in the future.

As I found out, it seems the order of "post engine init" is not garanteed, and it seems the AvaMediaEditor module is called before AvaMedia module's post engine init callback. The result is that the rundown server is started before some of the needed objects in ava media module are created. I worked around that by having some internal functions to ensure those objects (playback manager and managed instance cache) are created. The alternative would be to have a AvaMedia module have it's own PostInit event and AvaMediaEditor be a callback to that (might be more future proof?).

Also, as a precaution related to that, I added some functions to IAvaMediaModule to query the availability of the playback manager and instance cache so that shutdown sequences don't trigger check failures. I didn't get any, but it seems more prudent to have logic to avoid it in any case.

1.2 - Playback Context Management improvement:

It is intended for the rundown server and the rundow editor to work togheter (as well as possible) on the same rundown.
Example: stream deck can send some playback commands while the rundown editor is opened.
We want a rundown's playback context to only be closed when it is no longer needed by either editor or server.

The previous issue was if the rundown server would switch from one rundown to another, in the playback context, it would close the rundown's playback context, even if the editor was opened.
To fix that, a delegate has been added to query if the playback context can be closed. So that either the server or editor can keep it alive if needed and closed only when it no longer is.
This solution may be temporary until the refactor to have a separate rundown player (or playback context) object that can be either instanced exclusively or shared, which would be much more powerfull.

1.2.1 - New Playback Context Switch message

Adding a message to give feedback to the control app client about playback context switch. It is an important information to keep track of and having a dedicated message for it makes it more reliable.

Also added more fields to the FAvaRundownAssetsChanged to make it more usefull. I plan to add more event types in the future.

#jira UE-215154
#rb Juan.Portillo

[CL 33930994 by jean perrier in ue5-main branch]