[Backout] - CL33604992 [FYI] Dario.Mazzanti Original CL Desc ----------------------------------------------------------------- [Reference Viewer] Don't count redirectors in depth filter
RecursivelyPopulateNodeInfos now checks whether the current asset is actually a redirector. If that's the case, its Reference Node Info will be marked as such, and the Destination asset will be included in the graph.

RecursivelyCreateNodes will check for redirectors by using the flag set in Reference Node Info, and not increase the depth count in case of a redirector. Graph will explore the Destination asset as well.

#jira UE-199293
#rb aditya.ravichandran

[CL 33964475 by bob tellez in ue5-main branch]