Speed up symbol filtering in Insights
When loading a large symbol file into insights, the time spent filtering becomes very noticable. In total, it takes about 90 to load the symbols for a large binary. Of these 90s, about 72.5s are spent in the regex matcher in FResolvedSymbolFilter::Update. The regexes we are checking for could also be checked for by just scanning for a substring. With this change, the 73s shrink to < 300ms.
#rb ionut.matasaru

[CL 34028729 by sebastian schoner in ue5-main branch]