ChaosDebugDraw - fix race condition when creating a thread context for parallel-for
The race would happen if the owning thread is helping with tasks, which would be setting/unsetting the Frame pointer on that thread's context. We we reading from the parent thread context to get the Frame to use for the task thread on the task thread, which is not safe if the parent thread is running tasks. The solution is to copy the Frame pointer on the parent thread and pass it to the task, rather than passing the parent thread's context.

#jira UE-216178
#rb Sergio.Gardeazabal, vincent.robert

[CL 34042897 by chris caulfield in ue5-main branch]