Path Tracer: Add support for missing legacy shading models when using Substrate
Subsurface and two-sided foliage legacy shading models were not being handled properly.

Also refactor some of the sheen logic to be common among the various substrate shading models in the path tracer (so that the SheenQuality cvar applies the same to all).

Refactor substrate BSDFs to reduce the size of the temporary state needed during eval/sample. The energy terms E is only needed to setup probabilities, and the weight W is only needed during eval/sample. So we only need to keep the weight W in the temporary state. This optimization can be applied to non-substrate shading models as well, but can be done in a different CL.

#rb Sebastien.Hillaire

[CL 34079197 by chris kulla in ue5-main branch]