Implements experimental in-renderer lens distortion
Lens distortion is done solely in the TSR's HistoryUpdate pass for performance reason,
adjusting directly the alignment between rendered and history pixels. This
aproach has the advantage of not needing to change the resolution of the history
whenever the upscaling factor happening at the center of projection change due to
animated FOV or lens distortion, avoiding VMEM spikes as well as keeping the
TSR's history resurrection operational.

The lens distortion is implemented through a generic FLensDistortionLUT, that
can be generated conveniently with the panini reprojection FPaniniProjectionConfig
(configured with r.LensDistortion.Panini.D cvar) For this lens distortion to happen
in the history update perfectly, TSR requires both the FLensDistortionLUT's
distorting and the undistorting UV displacement LUTs.

Disclaimer:
- Any passes sampling pre-TSR-scene-texture after TSR doesn't correctly handle the lens distortion
yet (Motion Blur, Editor Selection outline, Editor primitives). This is deferred to
followup CL depending on the result.
- This will make post process material after TSR sampling scene textures even more
complicated, but this kind of effects already introduces aliasing and/or flickering anyway.
- The rendering resolution can be automatically adjusted to have 1 rendering pixel per display resolution
at ScreenPercentage=100 at the center of projection with r.LensDistortion.AffectScreenPercentage=1,
but this is intenionally disabled by default as this can reallocate renderer's internal resources whenever
the lens distortion or FOV is animated.

How to use:
0) Make sure you are in PIE, -game or a game build
1) Make sure the lens distortion support is compiled in TSR's shader on your platform with r.TSR.Support.LensDistortion
2) Make sure you are using TSR with r.AntiAliasingMethod=4
3) Make sure you have lens distortion in TSR enabled with r.TSR.LensDistortion=1 (can be toggled at runtime)
4) Generates FLensDistortionLUT with r.LensDistortion.Panini.D > 0 of your choosing
5) Make sure to disable motion blur with r.MotionBlurQuality=0 as it is not yet implemented

#jira UE-216401

[CL 34079251 by guillaume abadie in ue5-main branch]