Switchboard: Fix nDisplay node hitches that happen after SBL is left running for a few hours, by limiting NvApi calls that cause hitches to connction time instead of regularly.
We discovered that NvAPI calls can cause hitches in frame presentation, this CL adds fine control over what status data is polled and which should be explicitly requested by the user. This is used to request status data that does not change often (e.g. driver version) and also calls that have been determined to cause hitches when calling dxgi present, only at time of connection or when the users presses "Refresh Table", instead of regularly.

For the regular GPU stat calls, the NvAPI calls were replaced with NVMI calls, which do not seem to be causing hitches.

"Refresh Mosaics" button renamed to "Refresh Table" because it now refreshes more than just the mosaics. Now it also requests a full sync status poll. This intentional action might cause a hitch on the nodes but this is now triggered explicitly by the user, which is rarely needed anyway, instead of every second.

Also fixed an issue of the nDisplay monitor table not updating its display after emitting the dataChanged signal.

Fixed nDisplay Monitor not updating when the nDisplay node name or address changed.

Improved nDisplay monitor to show the connection icon instead of the wider column "Connected" with yes/no values.

Vertically aligned the cells in the table.

It now doesn't use the orange warning color on certain columns that are only relevant when the node is running (e.g. present mode).

Added a code-only option to disable exe flags button.

#rb zach.brockway
#jira UE-216292
#rn Fixed nDisplay hitches that happen after a few hours of running SBL.

[CL 34083929 by Alejandro Arango in ue5-main branch]