[ChaosVD] Initial implementation to add support to record and visualize game thread acceleration structures
This CL adds only AABB Tree are support

In addition as part of this change

- Added a way to generate a struct view that can be fed to the vanilla details panel for Solver Data structs. CVD solver data usually consist of a main struct that can contain several substructs in containers. When we select something in the viewport, we usually want to see the data in the parent struct and then some specific items in the containers.
Until now for these cases we created dedicated details panels with added functionality to interact with the data. But there are cases where we just want to see the struct data. Now for these cases we can just use this struct view API in the CVD Selection handle system to use the normal CVD's details panel.
- Added a way to allow CVD's main details panel to ignore struct types (needed to ignore structs that have their own data inspector widgets).
-Fixed an issue with CVD's Details panel where if you try to inspect a struct after uobject was inspected, the uobject details view was not being properly cleared out and hidden

#jira UE-216082, FORT-745804

#rb Benn.Gallagher

[CL 34084863 by sergio gardeazabal in ue5-main branch]