Report maximum bytes remaining for chunk-transferred content.
We use the bytes remaining to decide when to dispatch the next request. There are cases where we do not know the size of the data an activity is expecting; the HTTP message. During this phase a maximum size is reported. Another case is chunk-transferred content which is of an indefinite length - as we do not know how much data is left to receive, we report a very large amount.

#rb paul.chipchase
#rnx

[CL 34087290 by martin ridgers in ue5-main branch]