Fix crash when a view extension registers postprocessing pass callbacks for a disabled pass.
TOverridePassSequence allows the last enabled postprocessing pass to target the backbuffer directly. However, if a subsequent disabled pass has callbacks, those callbacks will run anyway and will try to use the backbuffer as an input, which is not supported (we crash with an assert in FRHITextureSRVCreateInfo::Validate because the backbuffer texture hasn't been created with TexCreate_ShaderResource). This change fixes the logic to take into account callbacks for disabled passes as well.

#rnx
#jira UE-194365
#ushell-cherrypick of 33765142 by mihnea.balta
#rb eric.renaudhoude, mihnea.balta, Ruslan.Idrisov

[CL 34107242 by zach bethel in ue5-main branch]