Iris - FieldOfViewNetObjectPrioritizer and tests.
The prioritizer consists of an inner sphere, outer sphere, a cone and a line of sight. Radii, field of view for cone, line of sight width and priorities are all configurable. Priorities cover the shape they're configured for except for the cone where the priority is linearly interpolated from the inner cone length to the full cone length. A 2D greyscale image is produced to memory by the VisualizeFieldOfViewNetObjectPrioritizer test.

#jira UE-216561
#rb Mattias.Hornlund

[CL 34123123 by peter engstrom in ue5-main branch]