[Chaos] Adding a null check to prevent a crash if a solver command function becomes unbound.
Also added ensures to try to track down the underlaying issue. The ensures should tell us if the issue is still happenning (without crashing), and if the sim callback object was created with an unbound function from the get go. (In case it is not, it means we could have a memeory stomp).

#rb Benn.Gallagher

[CL 34147122 by sergio gardeazabal in ue5-main branch]