Added AddRow variants to TEDS that provide an opportunity to initialize columns.
When adding a row to TEDS it's instantly created based on the provided table. This also causes observers to be immediately called which is either more expensive or causes subtle bugs. `BatchAddRow` already handles this through a callback and with this CR, `AddRow` now also has an option for that.

The following additional changes were made:
- All the table and row functions have had their use of `TypedElement*` updated to `TypedElementDataStorage::`.
- The implementation of `BatchAddRow` for reserved row handles no longer uses an intermediate array but instead uses the provided array through casting. This is a minor performance gain.

#jira UE-215244
#rb aditya.ravichandran
#rnx

[CL 34206501 by ronald koppers in ue5-main branch]