Add Default, First, Middle, and Last alignment to ToolMenu sections
The new alignment support layers on top of the existing insertion-order support. Sections are added to the toolbar in groups where we first add all Left-aligned sections according to their insertion order, then all First-aligned sections, and then the Middle and Last-aligned sections. Note that this can break the specified insertion order if the user mixed alignments, but otherwise it's maintained.

The First and Default alignments are grouped together, with First-aligned sections appearing before Default-aligned sections. The result is a toolbar with visual empty space between the First-Default group, Middle group, and the Last group. A caveat here is that the Middle group won't necessarily appear perfectly centered on the toolbar because it's position depends on the sizes of the First-Default and Last groups.

#jira UE-212286
#rb brooke.hubert, ross.smith2

[CL 34230831 by sebastian arleryd in ue5-main branch]