Fixed incorrect translation on ViewMatrix for shadow views, which could cause precision loss at LWC ranges, and fixed typo in CheckMatrixPrecision.
The ViewMatrix and OverriddenTranslatedViewMatrix were set to the same value, which is a matrix with 0 translation. When the view matrix is uploaded to GPU, it is first re-based to be relative to the ViewOrigin, before being truncated to single-precision. If the origin of the ViewMatrix is always 0, this rebasing process does not work correctly and precision loss can occur. This fix therefore applies a -ViewOrigin translation to the ViewMatrix, so has the correct translation in absolute worldspace.

#tests TM-VRSmoke (at 0,0,0 and 2T), TM-VSM (at 0 and 2T). Tested for changes in shadow view transform, and LOD selection of meshes
#rb Ola.Olsson

[CL 34231983 by wouter dek in ue5-main branch]