[ChaosVD] Fix incorrect context key used for the CVD UI Commands objects.
We were using the same key as the SActorPalette plugin by mistake. This results in an ensure followed of a crash when you have a actor palette widget open and try to use the Show menu in the viewport.

This is already fixed on UE5 main as part of a larger change in CVD made a month ago, but we were not aware it was causing a crash or other issues (I just spotted my mistake as part of that larger change and fixed it). This CL brings just the specific fix for UE 5.4.3.

#jira UE-217572
[REVIEW]
[FYI] Benn.Gallagher

[CL 34324685 by sergio gardeazabal in 5.4 branch]