Path Tracer: Change main radiance buffer to half precision
The goal is to save memory and enable the 16 bit version of the NNE denoiser. To maintain enough dynamic range, we have to apply the PreExposure factor during rendering. However, this factor can change during rendering, so we apply the difference when displaying the accumulated radiance buffer. In non-offline rendering cases, allow the base and current exposure factors to differ by at most 16x (4 stops) which should be sufficient to avoid issues in the case of auto-exposure.

Because we are now using exposure-relative accumulation, change the max path exposure into a relative control (and rename it to max path intensity). This is closer to the behavior of most other path tracers.

#rb Tiantian.Xie

[CL 34348064 by chris kulla in ue5-main branch]