DaySequenceModifierComponent: Convert modifier component to use player view position instead of player pawn position.
Modifier blend targets are now expected to be an APlayerController (or a class derived from it). We access the player camera manager to get the view position for the purposes of enabling/disabling and computing blend weight for the modifier.

Instead of relying on overlap events and tracking which pawn the player is currently viewing (which can change in replays, for example), we now poll the player view position every time the sequence player is updated. This solves many issues that were previously only solveable using game/project specific logic.

#rb lonnie.li
#jira UE-215543 UE-214888 UE-215138

#changelist validated
#virtualized

[CL 34382770 by nickolas drake in ue5-main branch]