[Motion Design] Display Cluster synchronization support for Level Streaming Playable.
* New Modular Feature for Synchronized Events. Simple interface to push events.

Synchronization for either clustering or forking becomes a first class citizen in the playable framework.
The current implementation uses the Display Cluster Events and is in a new plugin AvaDisplayCluster. More extension points will likely be added and implemented in that plugin.

There is also a "no sync" implementation in AvalancheMedia Module when running in non-clustered. I could eventually have another implementation for "forked" channels not based on nDisplay.

The implementation is selectable in Avalanche settings.

* Testing methodology:

The playback server a new test feature that will randomly delay the playback commands to cause desynchronization. It is enable with a cvar so it is possible to configure it from the command line or type it in the console.

Tracking the issues and desyncs required a lot of logging. I formatted the logs to be used with a log highlighter for notepad++ to make it easier to read. Combined with log replication, I could get a all on one log of all events on both client and servers in a good chronological order so I could track down when a desync occurs and what caused it.

* Playback Client Issues with desynchronized clustered/forked channels

As a result of these tests, I found a lot of weaknesses in the playback client side of things when dealing with "forked" channels. It required rewrites in remote transition and remote playable proxies to add tracking and reconcile of different states that can be desynchronized. The principle is, if it works when desynchronized, it will work when synchronized. But it has to work even if desynchronized. The client also has to be able to recover it's state, from the server(s), if it is shutdown and restarted and be able to continue operating the rundown without issues.

* Scene View Extension to Hide "Waiting" visible levels

Playable Groups have a scene view extension allowing playables to setup the view.
For the Level Streaming Playable, it is used to hide the level once it has become visible by the level streaming update, but should not yet be shown because it is waiting for other nodes in the cluster that are not yet ready.


#jira UE-215978
#rb Juan.Portillo

[CL 34399654 by jean perrier in ue5-main branch]